catchup

+++ travis-build.sh
@@ -15,7 +15,7 @@ if [ -n "${PHP_VERSION}" ]; then
     sed -i -- "s/PHP_VERSION=.*/PHP_VERSION=${PHP_VERSION}/g" .env
     sed -i -- 's/=false/=true/g' .env
     cat .env
-    docker-compose build
+    docker-compose build ${BUILD_SERVICE}
     docker images
 fi
 
@@ -30,7 +30,9 @@ if [ -n "${HUGO_VERSION}" ]; then
     mv ./${HUGO_BIN}/${HUGO_BIN} $HOME/bin/hugo
 
     # Remove existing docs
-    rm -r ./docs
+    if [ -d "./docs" ]; then
+        rm -r ./docs
+    fi
 
     # Build docs
     cd DOCUMENTATION
